Replace Old Build With New Simplified Build
