+++ .github/workflows/Mac-Build.yml
@@ -18,5 +18,3 @@ jobs:
         run: brew install gcc
       - name: Running build
         run: ruby scripts/build.rb
-      - name: Simplified build
-        run: ruby scripts/simple_build.rb
