+++ .github/workflows/Windows-Build.yml
@@ -16,5 +16,3 @@ jobs:
       - run: npm install -g yarn
       - name: Running build
         run: ruby scripts\build.rb
-      - name: Simplified build
-        run: ruby scripts/simple_build.rb
